+++ packages/adapters/src/adapters/Cornerstone3D/EllipticalROI.ts
@@ -1,13 +1,11 @@
 import { vec3 } from "gl-matrix";
 import { utilities } from "dcmjs";
-import { utilities as csUtilities } from "@cornerstonejs/core";
+
 import MeasurementReport from "./MeasurementReport";
 import BaseAdapter3D from "./BaseAdapter3D";
-
-type Point3 = [number, number, number];
+import { toScoord } from "../helpers";
 
 const { Ellipse: TID300Ellipse } = utilities.TID300;
-const { worldToImageCoords } = csUtilities;
 
 const EPSILON = 1e-4;
 
@@ -19,14 +17,11 @@ class EllipticalROI extends BaseAdapter3D {
     static getMeasurementData(
         MeasurementGroup,
         sopInstanceUIDToImageIdMap,
-        imageToWorldCoords,
         metadata
     ) {
         const {
             state,
             NUMGroup,
-            scoord,
-            scoordArgs,
             worldCoords,
             referencedImageId,
             ReferencedFrameNumber
@@ -37,68 +32,10 @@ class EllipticalROI extends BaseAdapter3D {
             EllipticalROI.toolType
         );
 
-        const [majorAxisStart, majorAxisEnd, minorAxisStart, minorAxisEnd] =
-            worldCoords;
-
-        const majorAxisVec = vec3.create();
-        vec3.sub(majorAxisVec, majorAxisEnd, majorAxisStart);
-
-        // normalize majorAxisVec to avoid scaling issues
-        vec3.normalize(majorAxisVec, majorAxisVec);
-
-        const minorAxisVec = vec3.create();
-        vec3.sub(minorAxisVec, minorAxisEnd, minorAxisStart);
-        vec3.normalize(minorAxisVec, minorAxisVec);
-
-        const imagePlaneModule = metadata.get(
-            "imagePlaneModule",
-            referencedImageId
-        );
-
-        if (!imagePlaneModule) {
-            throw new Error("imageId does not have imagePlaneModule metadata");
-        }
-
-        const { columnCosines } = imagePlaneModule;
-
-        // find which axis is parallel to the columnCosines
-        const columnCosinesVec = vec3.fromValues(
-            columnCosines[0],
-            columnCosines[1],
-            columnCosines[2]
-        );
-        const projectedMajorAxisOnColVec = vec3.dot(
-            columnCosinesVec,
-            majorAxisVec
-        );
-
-        const projectedMinorAxisOnColVec = vec3.dot(
-            columnCosinesVec,
-            minorAxisVec
-        );
-
-        const absoluteOfMajorDotProduct = Math.abs(projectedMajorAxisOnColVec);
-        const absoluteOfMinorDotProduct = Math.abs(projectedMinorAxisOnColVec);
-
-        let ellipsePoints = [];
-        if (Math.abs(absoluteOfMajorDotProduct - 1) < EPSILON) {
-            ellipsePoints = worldCoords;
-        } else if (Math.abs(absoluteOfMinorDotProduct - 1) < EPSILON) {
-            ellipsePoints = [
-                worldCoords[2],
-                worldCoords[3],
-                worldCoords[0],
-                worldCoords[1]
-            ];
-        } else {
-            console.warn("OBLIQUE ELLIPSE NOT YET SUPPORTED");
-            return null;
-        }
-
         state.annotation.data = {
             ...state.annotation.data,
             handles: {
-                points: [...ellipsePoints],
+                points: worldCoords,
                 activeHandleIndex: 0,
                 textBox: {
                     hasMoved: false
@@ -106,15 +43,15 @@ class EllipticalROI extends BaseAdapter3D {
             },
             frameNumber: ReferencedFrameNumber
         };
-        if (referencedImageId) {
-            state.annotation.data.cachedStats = {
-                [`imageId:${referencedImageId}`]: {
-                    area: NUMGroup
-                        ? NUMGroup.MeasuredValueSequence.NumericValue
-                        : 0
-                }
-            };
-        }
+        state.annotation.data.cachedStats = referencedImageId
+            ? {
+                  [`imageId:${referencedImageId}`]: {
+                      area: NUMGroup
+                          ? NUMGroup.MeasuredValueSequence.NumericValue
+                          : 0
+                  }
+              }
+            : {};
 
         return state;
     }
@@ -124,51 +61,55 @@ class EllipticalROI extends BaseAdapter3D {
         const { cachedStats = {}, handles } = data;
         const rotation = data.initialRotation || 0;
         const { referencedImageId } = metadata;
+        const scoordProps = {
+            is3DMeasurement,
+            referencedImageId
+        };
 
         let top, bottom, left, right;
 
-        // Using image coordinates for 2D points
-        // this way when it's restored we can assume the initial rotation is 0.
         if (rotation == 90 || rotation == 270) {
-            bottom = worldToImageCoords(referencedImageId, handles.points[2]);
-            top = worldToImageCoords(referencedImageId, handles.points[3]);
-            left = worldToImageCoords(referencedImageId, handles.points[0]);
-            right = worldToImageCoords(referencedImageId, handles.points[1]);
+            bottom = handles.points[2];
+            top = handles.points[3];
+            left = handles.points[0];
+            right = handles.points[1];
         } else {
-            top = worldToImageCoords(referencedImageId, handles.points[0]);
-            bottom = worldToImageCoords(referencedImageId, handles.points[1]);
-            left = worldToImageCoords(referencedImageId, handles.points[2]);
-            right = worldToImageCoords(referencedImageId, handles.points[3]);
+            top = handles.points[0];
+            bottom = handles.points[1];
+            left = handles.points[2];
+            right = handles.points[3];
         }
 
         // find the major axis and minor axis
-        const topBottomLength = Math.abs(top.y - bottom.y);
-        const leftRightLength = Math.abs(left.x - right.x);
+        const topBottomLength = Math.sqrt(
+            (top[0] - bottom[0]) ** 2 +
+                (top[1] - bottom[1]) ** 2 +
+                (top[2] - bottom[2]) ** 2
+        );
+        const leftRightLength = Math.sqrt(
+            (left[0] - right[0]) ** 2 +
+                (left[1] - right[1]) ** 2 +
+                (left[2] - right[2]) ** 2
+        );
 
         const points = [];
         if (topBottomLength > leftRightLength) {
             // major axis is bottom to top
-            points.push({ x: top[0], y: top[1] });
-            points.push({ x: bottom[0], y: bottom[1] });
-
-            // minor axis is left to right
-            points.push({ x: left[0], y: left[1] });
-            points.push({ x: right[0], y: right[1] });
+            points.push(top, bottom, left, right);
         } else {
             // major axis is left to right
-            points.push({ x: left[0], y: left[1] });
-            points.push({ x: right[0], y: right[1] });
-
-            // minor axis is bottom to top
-            points.push({ x: top[0], y: top[1] });
-            points.push({ x: bottom[0], y: bottom[1] });
+            points.push(left, right, top, bottom);
         }
 
         const { area } = cachedStats[`imageId:${referencedImageId}`] || {};
 
+        const convertedPoints = points.map(point =>
+            toScoord(scoordProps, point)
+        );
+
         return {
             area,
-            points,
+            points: convertedPoints,
             trackingIdentifierTextValue: this.trackingIdentifierTextValue,
             finding,
             findingSites: findingSites || [],