+++ packages/adapters/src/adapters/Cornerstone3D/UltrasoundDirectional.ts
@@ -12,38 +12,22 @@ class UltrasoundDirectional extends BaseAdapter3D {
     }
     // TODO: this function is required for all Cornerstone Tool Adapters, since it is called by MeasurementReport.
     static getMeasurementData(
-        MeasurementGroup,
+        measurementGroup,
         sopInstanceUIDToImageIdMap,
-        imageToWorldCoords,
         metadata
     ) {
-        const { defaultState, SCOORDGroup, ReferencedFrameNumber } =
+        const { state, worldCoords, ReferencedFrameNumber } =
             MeasurementReport.getSetupMeasurementData(
-                MeasurementGroup,
+                measurementGroup,
                 sopInstanceUIDToImageIdMap,
                 metadata,
-                UltrasoundDirectional.toolType
+                this.toolType
             );
-
-        const referencedImageId =
-            defaultState.annotation.metadata.referencedImageId;
-
-        const { GraphicData } = SCOORDGroup;
-        const worldCoords = [];
-        for (let i = 0; i < (GraphicData as number[]).length; i += 2) {
-            const point = imageToWorldCoords(referencedImageId, [
-                GraphicData[i],
-                GraphicData[i + 1]
-            ]);
-            worldCoords.push(point);
-        }
-
-        const state = defaultState;
 
         state.annotation.data = {
             ...state.annotation.data,
             handles: {
-                points: [worldCoords[0], worldCoords[1]],
+                points: worldCoords,
                 activeHandleIndex: 0,
                 textBox: {
                     hasMoved: false